Present output characters immediately before NL
dbluartsim was previously waiting for a newlin before flushing
its buffer to the screen and the TCP/IP port.  This had the
effect it so you couldn't ever see a prompt.  This version
fixes that, although a better fix is still needed--one that
perhaps waits before outputting a prompt until the UART is
idle for a period of time.